Add jquery logic to properly hide title.
No way to accurately calculate when the font was too big using aspect ratio alone (as resizing wasn’t proportional) so I have to do it in code instead.

Text needs to clip when overflown to properly tell when space has run out.

Using jquery as the simplest way to handle DOM manipulation.