Include Sass in single TypeScript entrypoint
We don't need to split the bundle into two entrypoints as we can import
the Sass code from the TypeScript. This means the build artifacts will
be in a single directory rather than split into ts and sass directories.
It also lets us give different names to the development and production
artefacts again.