build: bump Rust version
I bumped the Rust dependencies late last year, but I didn't realise that
an increase in MSRV is not considered a breaking change, so the Docker
build started failing because it was tagged to an old Rust compiler.
Bumping the compiler also broke as it will default to latest stable
version of Debian when it was cut, so the new image was using a newer
version of Debian with OpenSSL 3, which the old buster image doesn't
support. I should specify the Debian version in the builder image in the
future to avoid this compatibility issue. (I should also actually use
OpenSSL for reqwest as at the moment we aren't using TLS when connecting
to Last.fm...)