feat: add try_get_scrobble method
This allows us to immediately return cached scrobbles in the rendered
HTML if available, but never waste any time waiting for the last.fm API
to give us new data if the cache is stale. The 200ms timeout that was
previously in place was never long enough for the last.fm API to return
so it was just an unnecessary delay on response times.